requirements.txt
inquirer
termcolor
transformers
torch
Pillow
tensorflow